# backend/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: MeowHome API
  version: 0.2.0
  description: |
    猫宅 MeowHome 后端 API。移动优先的家庭多猫照护、记录、健康聚合与 AI 辅助。
    所有业务资源均通过 FamilyScope 授权，时间字段统一返回 RFC 3339（UTC 存储）。
servers:
  - url: /api/v1
    description: 默认前缀
...
paths:
...
  /auth/register:
    post:
      tags: [auth]
      summary: 注册
      operationId: authRegister
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags: [auth]
      summary: 登录
      operationId: authLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [auth]
      summary: 刷新令牌
      operationId: authRefresh
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: 刷新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [auth]
      summary: 登出
      operationId: authLogout
      responses:
        '204':
          description: 登出成功
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me:
    get:
      tags: [auth]
      summary: 当前用户
      operationId: authMe
      responses:
        '200':
          description: 当前用户信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /families:
    post:
      tags: [families]
      summary: 创建家庭
      operationId: createFamily
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FamilyCreateRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FamilyEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
  /families/{familyId}:
    get:
      tags: [families]
      summary: 获取家庭
      operationId: getFamily
      parameters:
        - $ref: '#/components/parameters/FamilyId'
      responses:
        '200':
          description: 家庭详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FamilyEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [families]
      summary: 更新家庭
      operationId: updateFamily
      parameters:
        - $ref: '#/components/parameters/FamilyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FamilyUpdateRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FamilyEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
  /families/{familyId}/members:
    get:
      tags: [families]
      summary: 成员列表
      operationId: listMembers
      parameters:
        - $ref: '#/components/parameters/FamilyId'
      responses:
        '200':
          description: 成员列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberListEnvelope'
  /families/{familyId}/cats:
    get:
      tags: [cats]
      summary: 猫咪列表
      operationId: listCats
      parameters:
        - $ref: '#/components/parameters/FamilyId'
      responses:
        '200':
          description: 猫咪列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatListEnvelope'
    post:
      tags: [cats]
      summary: 创建猫咪
      operationId: createCat
      parameters:
        - $ref: '#/components/parameters/FamilyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CatCreateRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatEnvelope'
  /families/{familyId}/cats/{catId}:
    get:
      tags: [cats]
      summary: 获取猫咪
      operationId: getCat
      parameters:
        - $ref: '#/components/parameters/FamilyId'
        - $ref: '#/components/parameters/CatId'
      responses:
        '200':
          description: 猫咪详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [cats]
      summary: 更新猫咪
      operationId: updateCat
      parameters:
        - $ref: '#/components/parameters/FamilyId'
        - $ref: '#/components/parameters/CatId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CatUpdateRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatEnvelope'
    delete:
      tags: [cats]
      summary: 删除猫咪（软删除）
      operationId: deleteCat
      parameters:
        - $ref: '#/components/parameters/FamilyId'
        - $ref: '#/components/parameters/CatId'
      responses:
        '204':
          description: 删除成功
  # --- Agent（猫管家）---
  /families/{familyId}/records/{recordId}:
    get:
      tags: [records]
      summary: 按编号读取本家庭未删除的原始记录
      operationId: getFamilyRecord
      parameters:
        - $ref: '#/components/parameters/FamilyId'
        - name: recordId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 原始记录详情，使用现有 Envelope
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: 记录不存在、已删除或不属于当前家庭
  /families/{familyId}/agent/messages:
    get:
      tags: [agent]
      summary: 家庭 Agent 消息列表（仅 visibility=family，不含私有聊天）
      description: |
        过滤不可识别的 type/status 时返回 400，不默认为全部数据。
        游标 before 为 `generated_at|id`，不允许跨家庭或绕过权限。
        仅返回 family 可见消息。默认 20 条，最多 50 条；next_cursor 始终存在，
        空字符串表示没有下一页。日期窗口按家庭 IANA 时区的本地日历计算，
        UTC 查询区间使用左闭右开边界。
      operationId: listAgentMessages
      parameters:
        - $ref: '#/components/parameters/FamilyId'
        - name: type
          in: query
          schema:
            type: string
            enum: [patrol_abnormal, patrol_weight, chat_answer, reminder_draft]
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, dismissed, expired]
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
        - name: before
          in: query
          schema: { type: string }
      responses:
        '200':
          description: 家庭消息列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentMessageListEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /families/{familyId}/agent/messages/{messageId}:
    get:
      tags: [agent]
      summary: Agent 消息详情
      operationId: getAgentMessage
      parameters:
        - $ref: '#/components/parameters/FamilyId'
        - name: messageId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 消息详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentMessageEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'
  /families/{familyId}/agent/messages/{messageId}/draft:
    patch:
      tags: [agent]
      summary: 创建或编辑该消息的唯一提醒草稿
      description: |
        未创建时 expected_version=0，首次保存后版本为 1；
        旧版本返回 AGENT_CONFLICT（409）。
      operationId: editAgentReminderDraft
      parameters:
        - $ref: '#/components/parameters/FamilyId'
        - name: messageId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentReminderEditRequest'
      responses:
        '200':
          description: 草稿已保存
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentReminderDraftEnvelope'
        '409':
          $ref: '#/components/responses/Conflict'
  /families/{familyId}/agent/messages/{messageId}/confirm:
    post:
      tags: [agent]
      summary: 确认提醒草稿（幂等）
      description: |
        服务端从消息行内加载草稿，客户端只提交 expected_version；
        再次校验身份、成员资格、可见性、猫咪归属、开关、草稿状态/有效期/计划时间。
        确认时草稿必须有未来的绝对 scheduled_at；只有文字时间的草稿可保存但不可确认。
        重复确认返回同一提醒，不新建。
      operationId: confirmAgentReminderDraft
      parameters:
        - $ref: '#/components/parameters/FamilyId'
        - name: messageId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentConfirmRequest'
      responses:
        '200':
          description: 确认成功（幂等）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentConfirmEnvelope'
        '409':
          $ref: '#/components/responses/Conflict'
        '403':
          $ref: '#/components/responses/Forbidden'
  /families/{familyId}/agent/messages/{messageId}/dismiss:
    post:
      tags: [agent]
      summary: 忽略 Agent 消息（幂等）
      description: 忽略仅隐藏消息，不撤销已创建的正式提醒。
      operationId: dismissAgentMessage
      parameters:
        - $ref: '#/components/parameters/FamilyId'
        - name: messageId
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: 忽略成功
        '403':
          $ref: '#/components/responses/Forbidden'
  /families/{familyId}/agent/chat:
    post:
      tags: [agent]
      summary: Agent 对话
      description: |
        首轮不携带 session_id 时按 family+user+client_message_id 恢复已创建会话。
        消息去除首尾空白后长度为 1–2000 个 Unicode 字符。每轮最多携带
        5 轮历史、4 次模型请求和 8 次工具执行，整体预算 25 秒。
        模型输出校验失败时返回 AGENT_INVALID_OUTPUT（422）；模型不可用时
        degraded=true，且兜底内容只能依据已成功取得的授权数据。
      operationId: agentChat
      parameters:
        - $ref: '#/components/parameters/FamilyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentChatRequest'
      responses:
        '200':
          description: 聊天结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentChatEnvelope'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          description: 模型输出不合法；Envelope.code 为 AGENT_INVALID_OUTPUT
        '503':
          description: Agent 关闭或模型不可用；Envelope.code 为 AGENT_DISABLED 或 AGENT_UNAVAILABLE
  /families/{familyId}/agent/sessions:
    get:
      tags: [agent]
      summary: 当前用户在本家庭的 Agent 会话列表
      operationId: listAgentSessions
      parameters:
        - $ref: '#/components/parameters/FamilyId'
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
        - name: before
          in: query
          schema: { type: string }
      responses:
        '200':
          description: 会话列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentSessionListEnvelope'
  /families/{familyId}/agent/sessions/{sessionId}/messages:
    get:
      tags: [agent]
      summary: 本人私有会话消息（游标分页）
      operationId: listAgentSessionMessages
      parameters:
        - $ref: '#/components/parameters/FamilyId'
        - name: sessionId
          in: path
          required: true
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
        - name: before
          in: query
          schema: { type: string }
      responses:
        '200':
          description: 会话消息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentSessionMessagesEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
  /internal/agent/patrol:
    post:
      tags: [agent]
      summary: 手动触发指定家庭巡检（仅目标家庭真实 admin，限流与审计）
      operationId: runAgentPatrol
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentPatrolRequest'
      responses:
        '200':
          description: 巡检执行结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentPatrolEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          description: RATE_LIMITED；每来源 IP 每分钟最多 3 次
components:
  parameters:
    FamilyId:
      name: familyId
      in: path
      required: true
      schema:
        type: string
        example: '01J'
    CatId:
      name: catId
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: 未认证或令牌失效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: 无权访问该家庭资源
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: 资源冲突（如邮箱已注册）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Envelope:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          example: OK
        message:
          type: string
        data:
          description: 业务数据
        request_id:
          type: string
    Error:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          example: VALIDATION_FAILED
        message:
          type: string
        request_id:
          type: string
    RegisterRequest:
      type: object
      required: [email, password, user_name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
        user_name:
          type: string
          minLength: 1
          maxLength: 64
        family_name:
          type: string
          description: 首次注册时创建的家庭名称
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    AuthResult:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          description: 访问令牌有效期（秒）
        user:
          $ref: '#/components/schemas/User'
    AuthEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AuthResult'
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        user_name:
          type: string
        family_id:
          type: string
        member_id:
          type: string
        role:
          type: string
          enum: [owner, admin, member]
        timezone:
          type: string
    UserEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/User'
    Family:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        timezone:
          type: string
        currency:
          type: string
        avatar_url:
          type: string
          nullable: true
    FamilyCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        timezone:
          type: string
          default: Asia/Shanghai
        currency:
          type: string
          default: CNY
    FamilyUpdateRequest:
      type: object
      properties:
        name:
          type: string
        timezone:
          type: string
        currency:
          type: string
        avatar_url:
          type: string
    FamilyEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Family'
    Member:
      type: object
      properties:
        id:
          type: string
        family_id:
          type: string
        user_id:
          type: string
        user_name:
          type: string
        role:
          type: string
          enum: [owner, admin, member]
    MemberListEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Member'
    Cat:
      type: object
      properties:
        id:
          type: string
        family_id:
          type: string
        name:
          type: string
        breed:
          type: string
          nullable: true
        gender:
          type: string
          enum: [male, female, unknown]
          nullable: true
        birth_date:
          type: string
          format: date
          nullable: true
        birthday:
          type: string
          format: date
          nullable: true
          description: Legacy client compatibility alias of birth_date.
        neutered:
          type: boolean
        diseases:
          type: array
          items:
            type: string
        allergies:
          type: array
          items:
            type: string
        color:
          type: string
          nullable: true
        avatar_url:
          type: string
          nullable: true
    CatCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        breed:
          type: string
        gender:
          type: string
          enum: [male, female, unknown]
        birth_date:
          type: string
          format: date
        neutered:
          type: boolean
        diseases:
          type: array
          items:
            type: string
        allergies:
          type: array
          items:
            type: string
        color:
          type: string
    CatUpdateRequest:
      type: object
      properties:
        name:
          type: string
        breed:
          type: string
        gender:
          type: string
          enum: [male, female, unknown]
        birth_date:
          type: string
          format: date
        neutered:
          type: boolean
        color:
          type: string
        avatar_key:
          type: string
    CatEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Cat'
    CatListEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Cat'
    # --- Agent（猫管家）schemas ---
    AgentMessage:
      type: object
      required: [id, role, visibility, type, severity, title, body, generated_at, model]
      properties:
        id: { type: string }
        session_id: { type: string, description: 对话消息所属会话；巡检消息可空 }
        role:
          type: string
          enum: [user, assistant, tool]
        visibility:
          type: string
          enum: [family, private]
        family_id: { type: string }
        user_id: { type: string, description: 系统消息可空 }
        cat_id: { type: string, description: 单猫消息的明确归属 }
        type:
          type: string
          enum: [patrol_abnormal, patrol_weight, chat_answer, reminder_draft]
        severity:
          type: string
          enum: [info, warning, danger]
        title: { type: string, maxLength: 200 }
        body: { type: string }
        evidence:
          type: array
          items:
            $ref: '#/components/schemas/AgentEvidence'
        action_suggestions:
          type: array
          items:
            $ref: '#/components/schemas/AgentAction'
        draft_version: { type: integer, minimum: 0 }
        draft_reminder:
          $ref: '#/components/schemas/AgentReminderInput'
        draft_expires_at: { type: string, format: date-time }
        action_status:
          type: string
          enum: [pending, confirmed, dismissed, expired]
        display_status:
          type: string
          enum: [dismissed]
          description: 忽略后从家庭消息列表隐藏；已确认提醒的 action_status 保持 confirmed
        generated_at: { type: string, format: date-time }
        model:
          type: string
          description: 规则引擎或实际 Provider 模型名；规则巡检为 rule-engine-v1
        disclaimer: { type: string }
    AgentEvidence:
      type: object
      required: [source_type, source_id]
      description: |
        R-03/R-05 的 source_type=reminder 且 source_id 为提醒 ID；R-04
        引用最近称重记录 ID，若从未称重则以 cat_profile 引用猫咪建档 ID；R-06 使用
        family_scope 和家庭 ID，并在 excerpt 写明查询的本地日期范围。
      properties:
        source_type:
          type: string
          enum: [record, reminder, trend, weight, cat_profile, family_scope]
        source_id: { type: string, description: 记录/提醒 ID；趋势类可空 }
        cat_ids:
          type: array
          items: { type: string }
        occurred_at: { type: string, format: date-time }
        excerpt: { type: string, description: 原始文案节选，禁止改写诊断 }
...
    AgentReminderInput:
      type: object
      required: [type, title]
      properties:
        cat_id: { type: string, description: 猫咪 ID 或 both；家庭级留空 }
        type:
          type: string
          enum: [medication, vaccine, deworm, checkup, weight, inventory, water, custom]
        title: { type: string, maxLength: 200 }
        subtitle: { type: string }
        time: { type: string, description: 展示用 time_label 兼容字段 }
        scheduled_at: { type: string, format: date-time, description: 可计算的绝对时间（UTC 存储） }
        timezone: { type: string, description: IANA 时区，展示用 }
        icon: { type: string }
        rule: { type: string }
    AgentReminderDraft:
      type: object
      required: [message_id, version, expires_at]
      properties:
        message_id: { type: string }
        version: { type: integer }
        expires_at: { type: string, format: date-time }
        reminder:
          $ref: '#/components/schemas/AgentReminderInput'
        confirmed_reminder_id: { type: string }
    AgentReminderResult:
      type: object
      required: [id, family_id, cat_id, type, title, state]
      properties:
        id: { type: string }
        family_id: { type: string }
        cat_id: { type: string }
        type: { type: string }
        title: { type: string }
        subtitle: { type: string }
        time: { type: string }
        state: { type: string }
        icon: { type: string }
        rule: { type: string }
        scheduled_at: { type: string, format: date-time }
        timezone: { type: string }
        completed_at: { type: string, format: date-time }
    AgentChatRequest:
      type: object
      required: [client_message_id, message]
      properties:
        session_id: { type: string, description: 首轮可缺省 }
        client_message_id: { type: string, minLength: 1, maxLength: 128, description: 幂等键，同会话内唯一 }
        message: { type: string, minLength: 1, maxLength: 2000 }
    AgentSessionSummary:
      type: object
      required: [id, created_at, updated_at]
      properties:
        id: { type: string }
        title: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    AgentReminderEditRequest:
      type: object
      required: [expected_version, reminder]
      properties:
        expected_version: { type: integer, minimum: 0 }
        reminder:
          $ref: '#/components/schemas/AgentReminderInput'
    AgentConfirmRequest:
      type: object
      required: [expected_version]
      properties:
        expected_version: { type: integer, minimum: 0 }
    AgentPatrolRequest:
      type: object
      required: [family_id]
      properties:
        family_id: { type: string }
    AgentMessageListEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              type: object
              required: [messages, next_cursor]
              properties:
                messages:
                  type: array
                  items:
                    $ref: '#/components/schemas/AgentMessage'
                next_cursor: { type: string }
    AgentMessageEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AgentMessage'
    AgentReminderDraftEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AgentReminderDraft'
    AgentChatEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              type: object
              required: [session_id, message, degraded]
              properties:
                session_id: { type: string }
                message:
                  $ref: '#/components/schemas/AgentMessage'
                degraded: { type: boolean }
    AgentConfirmEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              type: object
              required: [reminder, message_id, action_status]
              properties:
                reminder:
                  $ref: '#/components/schemas/AgentReminderResult'
                message_id: { type: string }
                action_status:
                  type: string
                  enum: [confirmed]
    AgentPatrolEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              type: object
              required: [messages]
              properties:
                messages:
                  type: array
                  items:
                    $ref: '#/components/schemas/AgentMessage'
    AgentSessionListEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              type: object
              required: [sessions, next_cursor]
              properties:
                sessions:
                  type: array
                  items:
                    $ref: '#/components/schemas/AgentSessionSummary'
                next_cursor: { type: string }
    AgentSessionMessagesEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              type: object
              required: [messages, next_cursor]
              properties:
                messages:
                  type: array
                  items:
                    $ref: '#/components/schemas/AgentMessage'
                next_cursor: { type: string }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []